+++ Classes/Configuration/FlexForms.php
@@ -30,6 +30,7 @@ class FlexForms
         'www',
         'dateOfBirth',
         'status',
+        'userGroup',
         'comments',
         'privacy',
     ];